+++ src/lib/wallets/connecters/binanceW3W/binanceW3W.ts
@@ -1,7 +1,7 @@
-import type { Connector } from "wagmi/connectors";
 import type { InjectedConnectorOptions } from "@wagmi/core/connectors/injected";
 import { InjectedConnector } from "wagmi/connectors/injected";
 import { Chain, Wallet, getWalletConnectConnector } from "@rainbow-me/rainbowkit";
+import { getWalletConnectUri, isAndroid } from "../helper";
 
 const isInBinance = () => {
   try {
@@ -11,40 +11,11 @@ const isInBinance = () => {
   }
 };
 
-export const getHref = (isAndroid: boolean, wc?: string) => {
-  const appID = "xoqXxUSMRccLCrZNRebmzj";
-  const startPagePath = "L3BhZ2VzL2Rhc2hib2FyZC1uZXcvaW5kZXg=";
-
-  let qs = `appId=${appID}&startPagePath=${startPagePath}`;
-  if (wc) {
-    const startPageQuery = encodeURI(`wc=${encodeURIComponent(wc)}&isDeepLink=true&id=${+new Date()}`);
-    qs = `${qs}&startPageQuery=${startPageQuery}`;
-  }
-  const host = "//app.binance.com";
-  if (isAndroid) {
-    return `bnc:${host}/mp/app?${qs}`;
-  }
-  return `https:${host}/?_dp=${encodeURI(`/mp/app?${qs}`)}`;
-};
-
 export interface BinanceW3WOptions {
   projectId: string;
   chains: Chain[];
   walletConnectVersion?: "2";
   walletConnectOptions?: any;
-}
-
-export async function getWalletConnectUri(connector: Connector, version: "1" | "2"): Promise<string> {
-  const provider = await connector.getProvider();
-  return version === "2"
-    ? new Promise<string>((resolve) => provider.once("display_uri", resolve))
-    : provider.connector.uri;
-}
-
-export function isAndroid(): boolean {
-  return (
-    typeof navigator !== "undefined" && /Android\s([0-9.]+)/.test(navigator.userAgent) // Source: https://github.com/DamonOehlman/detect-browser/blob/master/src/index.ts
-  );
 }
 
 export default function binanceW3W({
@@ -86,7 +57,7 @@ export default function binanceW3W({
       const getUri = async () => {
         const uri = await getWalletConnectUri(connector, walletConnectVersion);
 
-        return isAndroid() ? uri : `bitkeep://wc?uri=${encodeURIComponent(uri)}`;
+        return isAndroid() ? uri : `bnc://app.binance.com/mp/app?wc=${encodeURIComponent(uri)}`;
       };
 
       return {
