+++ src/lib/wallets/connecters/bitGet/bitGetWallet.ts
@@ -1,7 +1,7 @@
-import type { Connector } from "wagmi/connectors";
 import type { InjectedConnectorOptions } from "@wagmi/core/connectors/injected";
 import { InjectedConnector } from "wagmi/connectors/injected";
 import { Chain, Wallet, getWalletConnectConnector } from "@rainbow-me/rainbowkit";
+import { getWalletConnectUri, isAndroid } from "../helper";
 
 export interface BitKeepWalletLegacyOptions {
   projectId?: string;
@@ -15,19 +15,6 @@ export interface BitKeepWalletOptions {
   chains: Chain[];
   walletConnectVersion?: "2";
   walletConnectOptions?: any;
-}
-
-export async function getWalletConnectUri(connector: Connector, version: "1" | "2"): Promise<string> {
-  const provider = await connector.getProvider();
-  return version === "2"
-    ? new Promise<string>((resolve) => provider.once("display_uri", resolve))
-    : provider.connector.uri;
-}
-
-export function isAndroid(): boolean {
-  return (
-    typeof navigator !== "undefined" && /Android\s([0-9.]+)/.test(navigator.userAgent) // Source: https://github.com/DamonOehlman/detect-browser/blob/master/src/index.ts
-  );
 }
 
 export const bitGetWallet = ({
